+++ vendored/simple-graph/src/simple-graph.lisp
@@ -3,7 +3,7 @@
   (:export #:graph #:make-graph #:graph-nodes #:graph-edges
            #:add-node! #:add-edge!
            #:leaf? #:leaves
-           #:subgraph #:nodes-to #:paths-to
+           #:subgraph #:nodes-to #:paths-to #:flip-edges
            #:to-dot #:to-dot-with-stream)
   (:documentation "A very simple Graph library."))
 
@@ -11,8 +11,8 @@
 
 (defstruct graph
   "Hash Tables of nodes and edges."
-  (nodes (make-hash-table :test #'equal) :type hash-table)
-  (edges (make-hash-table :test #'equal) :type hash-table))
+  (nodes (make-hash-table :test #'equalp) :type hash-table)
+  (edges (make-hash-table :test #'equalp) :type hash-table))
 
 (defun add-node! (graph node)
   "Mutably add some NODE data to a GRAPH. Avoids adding the same node twice. Yields
@@ -70,6 +70,16 @@ same nodes can be added."
   (add-edge! g :a :c)
   (leaves g))
 
+(defun flip-edges (graph)
+  "Flip all the graph edges."
+  (let ((new (make-graph)))
+    (setf (graph-nodes new) (graph-nodes graph))
+    (maphash (lambda (node0 nodes0)
+               (dolist (node1 nodes0)
+                 (add-edge! new node1 node0)))
+             (graph-edges graph))
+    new))
+
 (defun subgraph (graph from &rest rest)
   "Form a new subgraph made up of all the nodes and edges reachable from some given
 starting nodes."
@@ -103,7 +113,7 @@ starting nodes."
     (labels ((recurse (acc)
                (multiple-value-bind (entry-p n edges) (iter)
                  (cond ((not entry-p) acc)
-                       ((member node edges) (recurse (cons n acc)))
+                       ((member node edges :test #'equalp) (recurse (cons n acc)))
                        (t (recurse acc))))))
       (recurse '()))))
 
@@ -143,21 +153,44 @@ starting nodes."
 
 (defun to-dot-with-stream (graph stream)
   "Write the GRAPH in dot format to some STREAM."
-  (format stream "graph {~%")
-  (maphash (lambda (node _)
-             (declare (ignore _))
-             (format stream "  \"~a\";~%" node))
-           (graph-nodes graph))
-  (maphash (lambda (node edges)
-             (dolist (edge edges)
-               (format stream "  \"~a\" -- \"~a\";~%" node edge)))
-           (graph-edges graph))
-  (format stream "}"))
+  (let ((curr 0)
+        (ixs (make-hash-table :test #'equalp)))
+    (format stream "graph {~%")
+    (maphash (lambda (node _)
+               (declare (ignore _))
+               (format stream "  ~a [label=\"~a\"];~%" curr node)
+               (setf (gethash node ixs) curr)
+               (incf curr))
+             (graph-nodes graph))
+    (maphash (lambda (node0 edges)
+               (let ((node0-id (gethash node0 ixs)))
+                 (dolist (node1 edges)
+                   (let ((node1-id (gethash node1 ixs)))
+                     (format stream "  ~a -- ~a;~%" node0-id node1-id)))))
+             (graph-edges graph))
+    (format stream "}")))
 
 (defun to-dot (graph)
   "Write the GRAPH to dot format as a string."
   (with-output-to-string (stream)
     (to-dot-with-stream graph stream)))
+
+#+nil
+(let ((g (make-graph)))
+  (add-node! g "A")
+  (add-node! g "A")
+  (add-node! g "A")
+  (add-node! g "B")
+  (add-node! g "C")
+  (add-node! g "D")
+  (add-node! g "E")
+  (add-node! g "F")
+  (add-edge! g "A" "B")
+  (add-edge! g "A" "C")
+  (add-edge! g "B" "D")
+  (add-edge! g "C" "D")
+  (add-edge! g "E" "F")
+  (to-dot g))
 
 #++
 (let ((g (make-graph)))
@@ -175,4 +208,4 @@ starting nodes."
   (add-edge! g "C" "D")
   (add-edge! g "E" "F")
   (with-open-file (stream #p"deps.dot" :direction :output :if-exists :supersede)
-    (simple-graph:to-dot-with-stream (subgraph g "A") stream)))
+    (to-dot-with-stream (subgraph g "A") stream)))