+++ vendored/transducers/transducers/package.lisp
@@ -19,8 +19,9 @@
   (:export #:branch #:inject #:split)
   ;; --- Reducers -- ;;
   (:export #:cons #:snoc #:vector #:string #:hash-table
-           #:count #:average #:median
+           #:count #:average #:median #:quantities
            #:any? #:all? #:anyp #:allp #:any #:all
+           #:partition
            #:first #:last
            #:fold #:max #:min #:find
            #:for #:for-each)