+++ vendored/filepaths/CHANGELOG.md
@@ -1,6 +1,12 @@
 # filepaths
 
-## Unreleased
+## 1.0.2 (2025-06-26)
+
+#### Added
+
+- Windows support, by accounting for the `:device` field of `pathname`.
+
+## 1.0.1 (2025-05-12)
 
 #### Fixed