+++ vendored/simple-graph/CHANGELOG.md
@@ -1,5 +1,16 @@
 # simple-graph
 
+## 1.1.0 (2025-05-15)
+
+#### Added
+
+- `flip-edges`
+
+#### Changed
+
+- Nodes are now compared via `equalp`.
+- DOT output has been improved.
+
 ## 1.0.0 (2025-01-13)
 
 Initial release.